assistant-platform/requirements.txt
# llama-cpp-python==0.3.16
numpy
pydantic>=2.0.0
gradio>=5.11.0
llama-index-core>=0.11.23
llama-index-embeddings-huggingface>=0.3.1
llama-index-vector-stores-chroma>=0.2.2
chromadb>=0.5.17
python-dotenv>=1.0.1
pyyaml>=6.0.2
colorama>=0.4.6
llama-index-llms-llama-cpp